finished work
scheme

* Implement a function findindex
Takes a list and an atom and returns a list of all indexes where the atom is present in the list. If the atom doesn’t exist, findindex should return -1.

*  Implement a function findrangeindex 
Takes a list, ℓ0, . . . , ℓn, a unary function f, and an atom x. It should return a list of list indices such that each item i in the output list satisfies f(ℓi) = x. 

*  Define a function intersection
Given two lists as input, return a list representing the set intersection of the lists. This means you need to convert lists to sets before applying any set operations. 

* Define a function compose
Defined to perform the same function as in slide 29 of the Subprogram lecture. That is, it should return a function that, when invoked and supplied an argument, will execute the function composition with the argument as input. For example, (compose inc inc)
should evaluate to #<procedure>, whereas (compose inc inc) 5 should evaluate to 7.

* Define function Map2
This function accepts a unary
function f and list l1, . . . , ln as inputs and evaluates to a new list f(l1), . . . , f(ln). Write a similar
function map2 which accepts a list j1, . . . , jn, another list ℓ1, . . . , ℓn (note they are of equal length),
a unary predicate p and a unary function f. It should evaluate to an n element list which, for all
1 ≤ i ≤ n, yields f(ℓi) if p(ji) holds, or ℓi otherwise. 

* Write a function skip which behaves as follows:
> ((skip 0) ’foo)
foo
> (((skip 1) ’foo) ’bar)
bar
> ((((skip 2) ’foo) ’bar) ’baz)
baz